You can't always cast pointer to int on because the sizes are different; use intptr_t

+++ src/gnu/usr.sbin/mkhybrid/src/match.c
@@ -24,6 +24,7 @@ static char rcsid[] ="$Id$";
 #include <stdlib.h>
 #endif
 #endif
+#include <inttypes.h>
 #include <string.h>
 #include "match.h"
 
@@ -114,9 +115,9 @@ char * fn;
   return 0; /* not found -> not excluded */
 }
 
-int i_ishidden()
+intptr_t i_ishidden()
 {
-  return((int)i_mat[0]);
+  return((intptr_t)i_mat[0]);
 }
 
 /* Joliet hide */
@@ -164,9 +165,9 @@ char * fn;
   return 0; /* not found -> not excluded */
 }
 
-int j_ishidden()
+intptr_t j_ishidden()
 {
-  return((int)j_mat[0]);
+  return((intptr_t)j_mat[0]);
 }
 
 #ifdef APPLE_HYB
@@ -234,9 +235,9 @@ char * fn;
   return 0; /* not found -> not excluded */
 }
 
-int hfs_ishidden()
+intptr_t hfs_ishidden()
 {
-  return((int)hfs_mat[0]);
+  return((intptr_t)hfs_mat[0]);
 }
 
 /* These will probably appear in mkisofs in the future */
